pyproject.toml
[tool.poetry]
name = "social-media-api"
version = "0.1.0"
description = "Python API development course"
authors = ["Devansh3712 <devanshamity@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.73.0"}
mypy = "^0.931"
pymongo = "^4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
